feat(style): add box-shadow style
