+++ crates/core/src/style/border.rs
@@ -1,0 +1,9 @@
+use crate::value::{BorderStyle, Colors, Unit};
+
+/// Border Style
+pub struct Border {
+    color: Colors,
+    radius: Unit,
+    width: Unit,
+    style: Vec<BorderStyle>,
+}
