+++ crates/core/src/style/mod.rs
@@ -1,6 +1,6 @@
 //! Evlis styles
-
 mod basic;
+mod border;
 mod bridge;
 mod column;
 mod flex;
@@ -10,6 +10,7 @@ mod typo;
 
 pub use self::{
     basic::{ContainerStyle, SizedBoxStyle},
+    border::Border,
     column::MultiColumnStyle,
     flex::FlexStyle,
     grid::GridStyle,
