+++ src/widgets/layouts/list.rs
@@ -1,0 +1,7 @@
+use elvis_core::Node;
+
+/// `List` is a set of poor orphan children, they don't have any style, just blowing in the wind.
+pub struct List {
+    /// List children
+    pub children: Vec<Node>,
+}
