+++ src/widgets/layouts/mod.rs
@@ -1,10 +1,12 @@
 //! Evlis layouts
-mod basic;
+mod r#box;
 mod column;
 mod flex;
 mod grid;
+mod list;
 
-pub use basic::{Container, List, SizedBox};
 pub use column::MultiColumn;
 pub use flex::{Align, Center, Col, Flex, Row};
 pub use grid::Grid;
+pub use list::List;
+pub use r#box::{Container, SizedBox};
